refactored compute script code to edge scripting code
