+++ src/EdgeScriptingAPI.php
@@ -4,8 +4,14 @@ declare(strict_types=1);
 
 namespace ToshY\BunnyNet;
 
+use Psr\Http\Client\ClientExceptionInterface;
 use ToshY\BunnyNet\Client\BunnyClient;
 use ToshY\BunnyNet\Enum\Host;
+use ToshY\BunnyNet\Helper\BodyContentHelper;
+use ToshY\BunnyNet\Model\API\EdgeScripting\Code\GetCode;
+use ToshY\BunnyNet\Model\API\EdgeScripting\Code\SetCode;
+use ToshY\BunnyNet\Model\Client\Interface\BunnyClientResponseInterface;
+use ToshY\BunnyNet\Validator\ParameterValidator;
 
 class EdgeScriptingAPI
 {
@@ -20,5 +26,48 @@ class EdgeScriptingAPI
         $this->client
             ->setApiKey($this->apiKey)
             ->setBaseUrl(Host::API_ENDPOINT);
+    }
+
+    /**
+     * @throws ClientExceptionInterface
+     * @throws Exception\BunnyClientResponseException
+     * @throws Exception\JSONException
+     * @return BunnyClientResponseInterface
+     * @param int $id
+     */
+    public function getCode(int $id): BunnyClientResponseInterface
+    {
+        $endpoint = new GetCode();
+
+        return $this->client->request(
+            endpoint: $endpoint,
+            parameters: [$id],
+        );
+    }
+
+    /**
+     * @throws ClientExceptionInterface
+     * @throws Exception\BunnyClientResponseException
+     * @throws Exception\JSONException
+     * @throws Exception\InvalidTypeForKeyValueException
+     * @throws Exception\InvalidTypeForListValueException
+     * @throws Exception\ParameterIsRequiredException
+     * @return BunnyClientResponseInterface
+     * @param int $id
+     * @param array<string,mixed> $body
+     */
+    public function setCode(
+        int $id,
+        array $body = [],
+    ): BunnyClientResponseInterface {
+        $endpoint = new SetCode();
+
+        ParameterValidator::validate($body, $endpoint->getBody());
+
+        return $this->client->request(
+            endpoint: $endpoint,
+            parameters: [$id],
+            body: BodyContentHelper::getBody($body),
+        );
     }
 }
