+++ src/Model/API/EdgeScripting/Code/SetCode.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace ToshY\BunnyNet\Model\API\Base\Compute;
+namespace ToshY\BunnyNet\Model\API\EdgeScripting\Code;
 
 use ToshY\BunnyNet\Enum\Header;
 use ToshY\BunnyNet\Enum\Method;
@@ -11,7 +11,7 @@ use ToshY\BunnyNet\Model\AbstractParameter;
 use ToshY\BunnyNet\Model\EndpointBodyInterface;
 use ToshY\BunnyNet\Model\EndpointInterface;
 
-class UpdateComputeScriptCode implements EndpointInterface, EndpointBodyInterface
+class SetCode implements EndpointInterface, EndpointBodyInterface
 {
     public function getMethod(): Method
     {
